fib test - fix undefined behavior warning found by clang
warning: passing an object that undergoes default argument promotion
to 'va_start' has undefined behavior [-Wvarargs]

+++ src/vnet/fib/fib_test.h
@@ -107,7 +107,7 @@ extern int fib_test_validate_lb_v(const load_balance_t *lb,
 
 extern int fib_test_validate_entry(fib_node_index_t fei,
                                    fib_forward_chain_type_t fct,
-                                   u16 n_buckets,
+                                   int n_buckets,
                                    ...);
 
 #endif
